fix: documentation update, rename divider with name to VoicesTextDivider
